citygroves/appgen/flesh/pyproject.toml
[tool.poetry]
name = "citygroves.appgen"
version = "0.1.0"
description = ""
authors = ["Damian Krystkiewicz <damian.krystkiewicz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0"
Pillow = "^6.2"
loguru = "^0.4.0"
graypy = "^2.1"
django-click = "^2.1"
psycopg2-binary = "^2.8"
djangorestframework = "^3.11"
celery = "^4.3"
django-environ = "^0.4.5"
redis = "^3.3"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
ipython = "^7.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
